+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Add NodeBluePrint.map
 - Add Doc class to create html documentations
 - Add Scope.reset and Node.reset
+- Add Node.addPlugin and removePlugin
 
 ### Changed